+++ fetcher/consumer.py
@@ -11,7 +11,7 @@ from huey.constants import WORKER_THREAD  # type: ignore
 from huey.consumer_options import ConsumerConfig  # type: ignore
 from prometheus_client import start_http_server
 
-from archiver.tasks import huey
+from fetcher.tasks import huey
 
 
 def main():
